02_Function_Tool_Calling/pyproject.toml
[project]
name = "02-tool-calling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.6.4",
    "python-dotenv>=1.2.1",
]
